requirements.txt
# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.10
langchain-groq>=0.2.0
langchain-openai>=0.1.0

# LLM Providers
groq>=0.5.0
openai>=1.0.0

# API and Search
serpapi>=0.1.5
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Geolocation
geopy>=2.4.0

# Web Scraping
selenium>=4.15.2
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
lxml>=4.9.0

# Web Frameworks
streamlit>=1.28.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
flask>=3.0.0
flask-cors>=4.0.0
asgiref>=3.7.0

# Data Validation
pydantic>=2.5.0
typing_extensions>=4.8.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
altair>=5.0.0
pillow>=10.0.0
